refactor: adjust API routes path
