+++ __tests__/integration/carts/updateCartItem.test.ts
@@ -2,7 +2,7 @@ import request from 'supertest';
 import app from '../../../src/app';
 import pool from '../../../src/config/db';
 
-describe('Delete cart item endpoint.', () => {
+describe('Update cart item endpoint.', () => {
   let userId: number;
   let token: string;
   let cartItemId: number;
@@ -20,7 +20,7 @@ describe('Delete cart item endpoint.', () => {
     token = res.body.data.accessToken;
     userId = res.body.data.user.id;
 
-    const cart = await request(app).post('/api/v1/cart')
+    const cart = await request(app).post('/api/v1/carts')
       .set('Authorization', `Bearer ${token}`)
       .send({ coffeeId: 1, quantity: 1, });
 
@@ -33,26 +33,26 @@ describe('Delete cart item endpoint.', () => {
   });
 
   it('Should return a 200 status code.', async () => {
-    const response = await request(app).delete('/api/v1/cart')
+    const response = await request(app).put('/api/v1/carts')
       .set('Authorization', `Bearer ${token}`)
-      .send({ cartItemId: cartItemId });
+      .send({ cartItemId: cartItemId, quantity: 1 });
 
     expect(response.statusCode).toBe(200);
     expect(response.body.status).toBe('OK');
   });
 
   it('Should return a 401 status code if access token is missing.', async () => {
-    const response = await request(app).delete('/api/v1/cart')
-      .send({ cartItemId: cartItemId });
+    const response = await request(app).put('/api/v1/carts')
+      .send({ cartItemId: cartItemId, quantity: 1 });
 
     expect(response.statusCode).toBe(401);
     expect(response.body.status).toBe('UNAUTHENTICATED');
   });
 
   it('Should return a 404 status code if cart item with provided id is not exist.', async () => {
-    const response = await request(app).delete('/api/v1/cart')
+    const response = await request(app).put('/api/v1/carts')
       .set('Authorization', `Bearer ${token}`)
-      .send({ cartItemId: -1 });
+      .send({ cartItemId: -1, quantity: 1 });
 
     expect(response.statusCode).toBe(404);
     expect(response.body.status).toBe('NOT_FOUND');
