+++ __tests__/integration/coffees/getAllCoffees.test.ts
@@ -2,7 +2,7 @@ import request from 'supertest';
 import app from '../../../src/app';
 import pool from '../../../src/config/db';
 
-describe('Get coffee details endpoint', () => {
+describe('Get all coffees endpoint', () => {
   let userId: number;
   let token: string;
   let coffeeId: number;
@@ -15,7 +15,7 @@ describe('Get coffee details endpoint', () => {
       'A classic coffee drink made by diluting a shot (or two) of espresso with hot water.',
       'https://example.com/americano.png',
     ]);
-
+    
     coffeeId = rows[0].id;
 
     const res = await request(app).post('/api/v1/auth/register')
@@ -37,27 +37,19 @@ describe('Get coffee details endpoint', () => {
     await pool.end();
   });
 
-  it('Should return a 200 status code and coffee details', async () => {
-    const response = await request(app).get(`/api/v1/coffee/${coffeeId}`)
+  it('Should return a 200 status code and list of coffees', async () => {
+    const response = await request(app).get('/api/v1/coffees')
       .set('Authorization', `Bearer ${token}`);
 
     expect(response.statusCode).toBe(200);
-    expect(response.body.data).toHaveProperty('coffee');
-    expect(response.body.data.coffee.name).toBe('Americano');
+    expect(response.body.data).toHaveProperty('coffees');
+    expect(Array.isArray(response.body.data.coffees)).toBe(true);
   });
-
+  
   it('Should return a 401 status code if access token is missing.', async () => {
-    const response = await request(app).get('/api/v1/coffee/1');
+    const response = await request(app).get('/api/v1/coffees');
 
     expect(response.statusCode).toBe(401);
     expect(response.body.status).toBe('UNAUTHENTICATED');
-  });
-
-  it('Should return a 404 status code if provided coffee id is not exist.', async () => {
-    const response = await request(app).get('/api/v1/coffee/-1')
-      .set('Authorization', `Bearer ${token}`);
-
-    expect(response.statusCode).toBe(404);
-    expect(response.body.status).toBe('NOT_FOUND');
   });
 });
