+++ __tests__/integration/orders/deleteOrder.test.ts
@@ -2,10 +2,10 @@ import request from 'supertest';
 import app from '../../../src/app';
 import pool from '../../../src/config/db';
 
-describe('Update cart item endpoint.', () => {
+describe('Delete order endpoint.', () => {
   let userId: number;
   let token: string;
-  let cartItemId: number;
+  let orderId: number;
 
   beforeAll(async () => {
     const res = await request(app).post('/api/v1/auth/register')
@@ -20,11 +20,14 @@ describe('Update cart item endpoint.', () => {
     token = res.body.data.accessToken;
     userId = res.body.data.user.id;
 
-    const cart = await request(app).post('/api/v1/cart')
+    await request(app).post('/api/v1/carts')
       .set('Authorization', `Bearer ${token}`)
       .send({ coffeeId: 1, quantity: 1, });
 
-    cartItemId = cart.body.data.cartItems[0].cart_item_id;
+    const order = await request(app).post('/api/v1/orders')
+      .set('Authorization', `Bearer ${token}`);
+
+    orderId = order.body.data.order.id;
   });
 
   afterAll(async () => {
@@ -32,27 +35,24 @@ describe('Update cart item endpoint.', () => {
     await pool.end();
   });
 
-  it('Should return a 200 status code.', async () => {
-    const response = await request(app).put('/api/v1/cart')
-      .set('Authorization', `Bearer ${token}`)
-      .send({ cartItemId: cartItemId, quantity: 1 });
+  it('Should return a 200 status code if order successfully deleted.', async () => {
+    const response = await request(app).delete(`/api/v1/orders/${orderId}`)
+      .set('Authorization', `Bearer ${token}`);
 
     expect(response.statusCode).toBe(200);
     expect(response.body.status).toBe('OK');
   });
 
   it('Should return a 401 status code if access token is missing.', async () => {
-    const response = await request(app).put('/api/v1/cart')
-      .send({ cartItemId: cartItemId, quantity: 1 });
+    const response = await request(app).delete(`/api/v1/orders/${orderId}`);
 
     expect(response.statusCode).toBe(401);
     expect(response.body.status).toBe('UNAUTHENTICATED');
   });
 
-  it('Should return a 404 status code if cart item with provided id is not exist.', async () => {
-    const response = await request(app).put('/api/v1/cart')
-      .set('Authorization', `Bearer ${token}`)
-      .send({ cartItemId: -1, quantity: 1 });
+  it('Should return a 404 status code if the cart with provided id is not exist.', async () => {
+    const response = await request(app).delete('/api/v1/orders/-1')
+      .set('Authorization', `Bearer ${token}`);
 
     expect(response.statusCode).toBe(404);
     expect(response.body.status).toBe('NOT_FOUND');
