+++ __tests__/integration/orders/getOrderDetails.test.ts
@@ -2,7 +2,7 @@ import request from 'supertest';
 import app from '../../../src/app';
 import pool from '../../../src/config/db';
 
-describe('Update order status endpoint.', () => {
+describe('Get order details endpoint.', () => {
   let userId: number;
   let token: string;
   let orderId: number;
@@ -20,13 +20,13 @@ describe('Update order status endpoint.', () => {
     token = res.body.data.accessToken;
     userId = res.body.data.user.id;
 
-    await request(app).post('/api/v1/cart')
+    await request(app).post('/api/v1/carts')
       .set('Authorization', `Bearer ${token}`)
       .send({ coffeeId: 1, quantity: 1, });
 
-    const order = await request(app).post('/api/v1/order')
+    const order = await request(app).post('/api/v1/orders')
       .set('Authorization', `Bearer ${token}`);
-
+    
     orderId = order.body.data.order.id;
   });
 
@@ -35,27 +35,26 @@ describe('Update order status endpoint.', () => {
     await pool.end();
   });
 
-  it('Should return a 200 status code if order status successfully updated.', async () => {
-    const response = await request(app).put(`/api/v1/order/${orderId}`)
-      .set('Authorization', `Bearer ${token}`)
-      .send({ status: 'paid' });
+  it('Should return a 200 status code if the order details successfully retrieved.', async () => {
+    const response = await request(app).get(`/api/v1/orders/${orderId}`)
+      .set('Authorization', `Bearer ${token}`);
 
     expect(response.statusCode).toBe(200);
     expect(response.body.status).toBe('OK');
+    expect(response.body.data).toHaveProperty('orders');
+    expect(Array.isArray(response.body.data.orders)).toBe(true);
   });
 
   it('Should return a 401 status code if access token is missing.', async () => {
-    const response = await request(app).put(`/api/v1/order/${orderId}`)
-      .send({ status: 'paid' });
+    const response = await request(app).get(`/api/v1/orders/${orderId}`);
 
     expect(response.statusCode).toBe(401);
     expect(response.body.status).toBe('UNAUTHENTICATED');
   });
 
   it('Should return a 404 status code if the cart with provided id is not exist.', async () => {
-    const response = await request(app).put('/api/v1/order/-1')
-      .set('Authorization', `Bearer ${token}`)
-      .send({ status: 'paid' });
+    const response = await request(app).get('/api/v1/orders/0')
+      .set('Authorization', `Bearer ${token}`);
 
     expect(response.statusCode).toBe(404);
     expect(response.body.status).toBe('NOT_FOUND');
