+++ __tests__/integration/orders/updateOrderStatus.test.ts
@@ -1,12 +1,11 @@
 import request from 'supertest';
 import app from '../../../src/app';
 import pool from '../../../src/config/db';
-import { generateAccessToken } from '../../../src/utilities/token';
 
-describe('Delete user account endpoint.', () => {
+describe('Update order status endpoint.', () => {
   let userId: number;
   let token: string;
-  let cookie: string;
+  let orderId: number;
 
   beforeAll(async () => {
     const res = await request(app).post('/api/v1/auth/register')
@@ -18,9 +17,17 @@ describe('Delete user account endpoint.', () => {
         address: 'Test street, Example, 00000'
       });
 
-    userId = res.body.data.user.id;
     token = res.body.data.accessToken;
-    cookie = res.headers['set-cookie'];
+    userId = res.body.data.user.id;
+
+    await request(app).post('/api/v1/carts')
+      .set('Authorization', `Bearer ${token}`)
+      .send({ coffeeId: 1, quantity: 1, });
+
+    const order = await request(app).post('/api/v1/orders')
+      .set('Authorization', `Bearer ${token}`);
+
+    orderId = order.body.data.order.id;
   });
 
   afterAll(async () => {
@@ -28,28 +35,27 @@ describe('Delete user account endpoint.', () => {
     await pool.end();
   });
 
-  it('Should return a 200 status code if user successfully deleted', async () => {
-    const response = await request(app).delete('/api/v1/user/delete')
+  it('Should return a 200 status code if order status successfully updated.', async () => {
+    const response = await request(app).put(`/api/v1/orders/${orderId}`)
       .set('Authorization', `Bearer ${token}`)
-      .set('Cookie', cookie);
+      .send({ status: 'paid' });
 
     expect(response.statusCode).toBe(200);
-    expect(response.body.message).toBe('User has been deleted!');
+    expect(response.body.status).toBe('OK');
   });
 
   it('Should return a 401 status code if access token is missing.', async () => {
-    const response = await request(app).delete('/api/v1/user/delete')
-      .set('Cookie', cookie);
+    const response = await request(app).put(`/api/v1/orders/${orderId}`)
+      .send({ status: 'paid' });
 
     expect(response.statusCode).toBe(401);
     expect(response.body.status).toBe('UNAUTHENTICATED');
   });
 
-  it('Should return a 401 status code if access token is missing.', async () => {
-    const fakeToken = generateAccessToken(-1);
-    const response = await request(app).delete('/api/v1/user/delete')
-      .set('Authorization', `Bearer ${fakeToken}`)
-      .set('Cookie', cookie);
+  it('Should return a 404 status code if the cart with provided id is not exist.', async () => {
+    const response = await request(app).put('/api/v1/orders/-1')
+      .set('Authorization', `Bearer ${token}`)
+      .send({ status: 'paid' });
 
     expect(response.statusCode).toBe(404);
     expect(response.body.status).toBe('NOT_FOUND');
