+++ __tests__/integration/users/deleteUser.test.ts
@@ -1,10 +1,12 @@
 import request from 'supertest';
 import app from '../../../src/app';
 import pool from '../../../src/config/db';
+import { generateAccessToken } from '../../../src/utilities/token';
 
-describe('Create order endpoint.', () => {
+describe('Delete user account endpoint.', () => {
   let userId: number;
   let token: string;
+  let cookie: string;
 
   beforeAll(async () => {
     const res = await request(app).post('/api/v1/auth/register')
@@ -16,8 +18,9 @@ describe('Create order endpoint.', () => {
         address: 'Test street, Example, 00000'
       });
 
-    token = res.body.data.accessToken;
     userId = res.body.data.user.id;
+    token = res.body.data.accessToken;
+    cookie = res.headers['set-cookie'];
   });
 
   afterAll(async () => {
@@ -25,32 +28,30 @@ describe('Create order endpoint.', () => {
     await pool.end();
   });
 
-  it('Should return a 201 status code if order successfully created.', async () => {
-    await request(app).post('/api/v1/cart')
+  it('Should return a 200 status code if user successfully deleted', async () => {
+    const response = await request(app).delete('/api/v1/users/delete')
       .set('Authorization', `Bearer ${token}`)
-      .send({ coffeeId: 1, quantity: 1, });
+      .set('Cookie', cookie);
 
-    const response = await request(app).post('/api/v1/order')
-      .set('Authorization', `Bearer ${token}`);
-
-    expect(response.statusCode).toBe(201);
-    expect(response.body.status).toBe('CREATED');
-    expect(response.body.data).toHaveProperty('order');
+    expect(response.statusCode).toBe(200);
+    expect(response.body.message).toBe('User has been deleted!');
   });
 
   it('Should return a 401 status code if access token is missing.', async () => {
-    const response = await request(app).post('/api/v1/order');
+    const response = await request(app).delete('/api/v1/users/delete')
+      .set('Cookie', cookie);
 
     expect(response.statusCode).toBe(401);
     expect(response.body.status).toBe('UNAUTHENTICATED');
   });
 
-  it('Should return a 404 status code if the cart with provided id is not exist.', async () => {
-    const response = await request(app).post('/api/v1/order')
-      .set('Authorization', `Bearer ${token}`);
+  it('Should return a 401 status code if access token is missing.', async () => {
+    const fakeToken = generateAccessToken(-1);
+    const response = await request(app).delete('/api/v1/users/delete')
+      .set('Authorization', `Bearer ${fakeToken}`)
+      .set('Cookie', cookie);
 
     expect(response.statusCode).toBe(404);
     expect(response.body.status).toBe('NOT_FOUND');
-
   });
 });
